requirements.txt
streamlit==1.38.0
pillow>=10.4.0
opencv-python>=4.10.0.84
numpy>=1.26.0
pandas>=2.2.2
pydantic>=2.7.4
rapidfuzz>=3.9.6
jiwer>=3.0.4
pytesseract>=0.3.10
paddleocr>=2.7.0
paddlepaddle==2.6.1
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.44.0
sentencepiece>=0.2.0
timm>=1.0.7
scikit-image>=0.24.0
python-dotenv>=1.0.1
openai>=1.40.0
pyyaml>=6.0.2
pyyaml-include>=1.4.1
pdf2image>=1.17.0
PyPDF2>=3.0.1
PyMuPDF>=1.24.9

